% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_variables.R
\name{append_rowwise_summary_indicator}
\alias{append_rowwise_summary_indicator}
\title{Append a rowwise summary indicator column}
\usage{
append_rowwise_summary_indicator(
  df,
  cols,
  yes_values,
  new_indicator_colname,
  new_indicator_var_label,
  new_indicator_yes_value = "Yes",
  new_indicator_no_value = "No"
)
}
\arguments{
\item{df}{Data frame}

\item{cols}{Character vector of columns to summarise}

\item{yes_values}{Vector. \code{new_indicator_colname} will be "Yes" if any
values in \code{cols} (rowwise) match one of these values.}

\item{new_indicator_colname}{Name of new indicator column.}

\item{new_indicator_var_label}{Character string to set as variable label for
new indicator column.}

\item{new_indicator_yes_value}{Default is "Yes".}

\item{new_indicator_no_value}{Default is "No"}
}
\value{
A data frame with an additional column, named as per
\code{new_indicator_colname}.
}
\description{
Low level function that appends a (binary) indicator column to a dataframe,
summarising rowwise whether any values in a selection of existing columns
match one or more specified values.
}
\examples{
# example input df
df <- tibble::tribble(
~med_1, ~med_2, ~med_3, ~IGNORED_COL,
"Beta blocker", NA, NA, "Beta blocker",
"ACE inhibitor", "Beta blocker", NA, "Beta blocker",
"Aspirin", "Aspirin", NA, "Beta blocker",
"Aspirin", "ACE inhibitor", NA, "Beta blocker"
)

# add indicator col
df <- append_rowwise_summary_indicator(df = df,
cols = c("med_1", "med_2", "med_3"),
yes_values = c("Beta blocker", "ACE inhibitor"),
new_indicator_colname = "on_anti_hypertensive",
new_indicator_var_label = "Taking anti-hypertensive medication",
new_indicator_yes_value = "Yes",
new_indicator_no_value = "No")

# print result
print(df)

# print variable label for new column
print(attributes(df$on_anti_hypertensive)$label)
}
