% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_control_match.R
\name{create_case_control_matched_df}
\alias{create_case_control_matched_df}
\title{Create a data frame with cases and matched controls}
\usage{
create_case_control_matched_df(
  df,
  case_control_colname,
  case_value = "Case",
  control_value = "Control",
  matching_vars_categorical = NULL,
  matching_vars_continuous = NULL,
  n_control_ratio = 4,
  seed = 42
)
}
\arguments{
\item{df}{A data frame}

\item{case_control_colname}{Character. The name of an indicator column in
\code{df} that labels individuals as either cases or controls.}

\item{case_value}{String Used to determine which individuals are cases. This
value should be present in \code{df[[case_control_colname]]}.}

\item{control_value}{String Used to determine which individuals are controls
This value should be present in \code{df[[case_control_colname]]}.}

\item{matching_vars_categorical}{Character vector of column names containing
categorical variables to match on. These columns should all be of type
character.}

\item{matching_vars_continuous}{Named integer vector. Names should be column
names containing continuous variables to match on (these columns should all
be of type numeric). Integer values are the number of digits to match to.
For example, \code{c("age_yrs" = 0)} specifies that controls will be matched on
'age_yrs' to the nearest year.}

\item{n_control_ratio}{Integer. The number of controls to match for each
case.}

\item{seed}{Integer, for reproducibility.}
}
\value{
A data frame
}
\description{
Create a data frame with cases and matched controls
}
