% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_ukb.R
\name{create_unique_id_df}
\alias{create_unique_id_df}
\title{Create a UKB data frame with a unique ID column}
\usage{
create_unique_id_df(
  path,
  delim = "\\t",
  ukb_data_dict = ukbwranglr::get_ukb_data_dict(),
  ukb_codings = ukbwranglr::get_ukb_codings(),
  descriptive_colnames = TRUE,
  label = FALSE,
  field_ids = c("31", "52", "34", "21000", "53", "96", "50"),
  instances = "0",
  id_col = "..unique_id",
  remove = TRUE,
  .ignore_duplicate_ids = FALSE
)
}
\arguments{
\item{path}{The path to a UK Biobank main dataset file.}

\item{delim}{Delimiter for the UKB main dataset file. Default is "auto" (see
\code{\link[data.table:fread]{data.table::fread()}}). Ignored if the file name ends with \code{.dta}
(i.e. is a STATA file) or if \code{ukb_main} is a data frame.}

\item{ukb_data_dict}{The UKB data dictionary (available online at the UK
Biobank
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{data
showcase}. This should be a data frame where all columns are of type
\code{character}.}

\item{ukb_codings}{The UKB codings file (available online at the UK Biobank
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{data
showcase}. This should be a data frame where all columns are of type
\code{character}.}

\item{descriptive_colnames}{If \code{TRUE}, rename columns with longer
descriptive names derived from the UK Biobank's data dictionary 'Field'
column.}

\item{label}{If \code{TRUE}, apply variable labels and label coded values as
factors.}

\item{field_ids}{A character vector of fields IDs that will be used to create
the new unique ID column. These should match the values under column
'Field' in the UK Biobank data dictionary.}

\item{instances}{A character vector of instances to include when generating
the new unique ID column. Should contain one or more of the following
digits: '0', '1', '2', '3'. Note that more recent datasets may include
instances that are not present in older datasets. By default only the first
instance is used.}

\item{id_col}{Name of the the new column to be created.}

\item{remove}{If \code{TRUE}, remove input columns from output data frame.}

\item{.ignore_duplicate_ids}{If \code{TRUE}, allow duplicate ID values and
raise a warning if any are found. May be helpful for debugging. By default
this is \code{FALSE}.}
}
\value{
A data frame
}
\description{
A convenience function that returns a data frame for a main UK Biobank
dataset with a unique ID column, created by concatenating values from a
selection of variables. \strong{Manual validation of any subsequent linkage is
strongly advised.}
}
\seealso{
\code{\link{create_unique_id}}
}
