% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliber.R
\name{default_overlapping_disease_categories_csv}
\alias{default_overlapping_disease_categories_csv}
\title{CSV file for resolving overlapping disease categories in mapped CALIBER codes}
\usage{
default_overlapping_disease_categories_csv()
}
\value{
A file path.
}
\description{
Returns the file path to csv included with this package for resolving
overlapping disease categories in mapped CALIBER codes. To be used with
\code{\link[=reformat_caliber_for_ukb]{reformat_caliber_for_ukb()}}.
}
\examples{
# return file path
default_overlapping_disease_categories_csv()

# read file into R
read.csv(default_overlapping_disease_categories_csv())
}
\seealso{
Other CALIBER: 
\code{\link{download_caliber_repo}()},
\code{\link{get_caliber_categories_mapping}()},
\code{\link{read_caliber_raw}()},
\code{\link{reformat_caliber_for_ukb}()}
}
\concept{CALIBER}
