% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_variables.R
\name{derive_touchscreen_self_reported_med_bp_dm_chol_exog_hormones}
\alias{derive_touchscreen_self_reported_med_bp_dm_chol_exog_hormones}
\title{Summarise UK Biobank touchscreen responses for self-reported medications
(cholesterol, blood pressure, diabetes or take exogenous hormones)}
\usage{
derive_touchscreen_self_reported_med_bp_dm_chol_exog_hormones(
  ukb_main,
  ukb_data_dict = ukbwranglr::get_ukb_data_dict(),
  instance = NULL,
  .drop = TRUE
)
}
\arguments{
\item{ukb_data_dict}{The UKB data dictionary (available online at the UK
Biobank
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{data
 showcase}. This should be a data frame where all columns are of type
\code{character}. By default, this is downloaded from the
\href{https://github.com/rmgpanw/ukbwranglr_resources}{\code{ukbwranglr_resources}}
 github repo.}

\item{instance}{Character. Optionally specify a single instance to be
summarised e.g. '1'. Default is \code{NULL}, in which case the appended columns
will reflect a summary across all available instances.}

\item{.drop}{If \code{TRUE}, remove input columns for Field IDs 6153 and 6177 from
output.}
}
\value{
A data frame.
}
\description{
Summarises field IDs
\href{https://biobank.ndph.ox.ac.uk/ukb/field.cgi?id=6153}{6153} and
\href{https://biobank.ndph.ox.ac.uk/ukb/field.cgi?id=6177}{6177}. The returned
data frame contains separate indicator columns for each medication type.
}
\details{
Note that Field ID 6153 applies only to female participants, where as Field
ID applies to male participants.
}
