% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliber.R
\name{download_caliber_repo}
\alias{download_caliber_repo}
\title{Download the caliber github repository}
\usage{
download_caliber_repo(
 
    url = "https://github.com/spiros/chronological-map-phenotypes/archive/refs/heads/master.zip"
)
}
\arguments{
\item{url}{}
}
\value{
File path to downloaded (and unzipped) repository.
}
\description{
Downloads to \code{tempdir()} and unzips.
}
