% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy_caliber_dir_path}
\alias{dummy_caliber_dir_path}
\title{Dummy CALIBER repository}
\usage{
dummy_caliber_dir_path()
}
\value{
A string.
}
\description{
Returns the file path to a dummy data
\href{https://github.com/spiros/chronological-map-phenotypes}{CALIBER} repository.
}
\concept{Dummy data}
