% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dummy_manhattan_df}
\alias{dummy_manhattan_df}
\title{Dummy data frame for \code{plot_manhattan()}}
\usage{
dummy_manhattan_df(type = "gwas")
}
\arguments{
\item{type}{Either 'gwas' or 'phewas}
}
\value{
A data frame
}
\description{
Dummy data frame for \code{plot_manhattan()}
}
\examples{
dummy_manhattan_df()
}
