% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliber.R
\name{get_caliber_categories_mapping}
\alias{get_caliber_categories_mapping}
\title{Get a mapping table of CALIBER phenotypes/categories}
\usage{
get_caliber_categories_mapping()
}
\value{
A data frame
}
\description{
Includes a category order and colour for plotting. These were manually
assigned based on those used in the \link[PheWAS:phewas-package]{PheWAS::PheWAS} package (see
\link[PheWAS:pheinfo]{PheWAS::pheinfo}). The original CALIBER phenotype to categories mapping file
was obtained from CALIBER's \href{https://github.com/spiros/chronological-map-phenotypes/blob/master/phenotype_categories_mapping.csv}{github repo}.
}
\examples{
get_caliber_categories_mapping()
}
\seealso{
Other CALIBER: 
\code{\link{default_overlapping_disease_categories_csv}()},
\code{\link{download_caliber_repo}()},
\code{\link{read_caliber_raw}()},
\code{\link{reformat_caliber_for_ukb}()}
}
\concept{CALIBER}
