% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukb_clinical_events.R
\name{get_clinical_events_source}
\alias{get_clinical_events_source}
\title{Filter UK Biobank clinical events for selected data sources}
\usage{
get_clinical_events_source(
  clinical_events,
  sources,
  allow_missing_sources = FALSE
)
}
\arguments{
\item{clinical_events}{A clinical events table (data frame or \code{tbl_dbi}
object) created by \code{\link[ukbwranglr:tidy_clinical_events]{ukbwranglr::tidy_clinical_events()}}.}

\item{sources}{A character vector of data sources. Must be listed under
\code{source} in \code{\link[ukbwranglr:clinical_events_sources]{ukbwranglr::clinical_events_sources()}}.}

\item{allow_missing_sources}{If \code{FALSE} (default), an error is raised if any
values for \code{sources} are not present in \code{clinical_events}. If \code{TRUE}, a
warning is raised instead.}
}
\value{
A data frame
}
\description{
Filter UK Biobank clinical events for selected data sources
}
