% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukb_synthetic_dataset.R
\name{get_ukb_synthetic_dataset_html_links}
\alias{get_ukb_synthetic_dataset_html_links}
\title{Get URL links to from UKB synthetic dataset webpage}
\usage{
get_ukb_synthetic_dataset_html_links(pattern = "md5$")
}
\arguments{
\item{pattern}{Optionally filter the list of URLs for a subset matching a
pattern.}
}
\value{
A character vector of URLs.
}
\description{
Get URL links to from UKB synthetic dataset webpage
}
\examples{
\dontrun{
get_ukb_synthetic_dataset_html_links()
}
}
