% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliber.R
\name{map_caliber}
\alias{map_caliber}
\title{Map CALIBER codelist data frame from one coding system to another}
\usage{
map_caliber(df, from, to, all_lkps_maps, col_filters)
}
\arguments{
\item{df}{Reformatted CALIBER code list, data frame}

\item{from}{Code type to map from}

\item{to}{Code type to map to}

\item{all_lkps_maps}{Named list}

\item{col_filters}{See \code{\link[codemapper:map_codes]{codemapper::map_codes()}}}
}
\value{
A data frame. The \code{code_type} column will equal the value for \code{to}.
}
\description{
Helper function for \code{\link[=reformat_caliber_for_ukb]{reformat_caliber_for_ukb()}}. Maps a codes in a
reformatted CALIEBR codelist data frame, used for Read 2 to Read 3, and ICD10
to ICD9.
}
