% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliber.R
\name{map_caliber_for_ukb}
\alias{map_caliber_for_ukb}
\title{Reformat and map CALIBER codes for use with UK Biobank data}
\usage{
map_caliber_for_ukb(caliber_dir_path = download_caliber_repo(), all_lkps_maps)
}
\arguments{
\item{caliber_dir_path}{Path to CALIBER repository directory.}

\item{all_lkps_maps}{Either a named list of lookup and mapping tables (either
data frames or \code{tbl_dbi} objects), or the path to a SQLite database
containing these tables (see also \code{\link[codemapper:build_all_lkps_maps]{build_all_lkps_maps()}} and
\code{\link[codemapper:all_lkps_maps_to_db]{all_lkps_maps_to_db()}}). By default, will look for a SQLite datbase called
\code{all_lkps_maps.db} in the current working directory.}
}
\value{
A data frame.
}
\description{
Reformat and map CALIBER codes for use with UK Biobank data
}
\seealso{
Other CALIBER: 
\code{\link{download_caliber_repo}()},
\code{\link{get_caliber_categories_mapping}()}
}
\concept{CALIBER}
