% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliber.R
\name{read_caliber_raw}
\alias{read_caliber_raw}
\title{Read a local copy of the CALIBER repository into R}
\usage{
read_caliber_raw(caliber_dir_path, overlapping_disease_categories = "error")
}
\arguments{
\item{caliber_dir_path}{Path to a locally downloaded copy of the \href{https://github.com/spiros/chronological-map-phenotypes}{CALIBER github repository}.}

\item{overlapping_disease_categories}{If 'error' (default), raises an error
if any overlapping disease categories are present after mapping. Specify
'warning' to raise a warning instead.}
}
\value{
A named list of data frames.
}
\description{
All CALIBER codes are read into a named list containing 3 data
frames: primary care Read 2, secondary care ICD10 and secondary care OPCS4
codes.
}
\details{
The directory supplied to \code{caliber_dir_path} is expected to contain
subdirectories \code{primary_care} and \code{secondary_care}, each of which contains
csv files with clinical code lists.

Note also that:
\enumerate{
\item Medcodes are dropped
\item Minimal reformatting is performed so that primary and secondary care
codes may be combined into a single data frame
}
}
\examples{
# read local copy of CALIBER repository into a named list. Note that
# (i) Medcodes are dropped and (ii) minimal reformatting is performed so that
# primary and secondary care codes may be combined into a single data frame
caliber_raw <- read_caliber_raw(dummy_caliber_dir_path())
caliber_raw

# combine into a single data frame with dplyr
dplyr::bind_rows(caliber_raw)
}
\seealso{
Other CALIBER: 
\code{\link{default_overlapping_disease_categories_csv}()},
\code{\link{download_caliber_repo}()},
\code{\link{get_caliber_categories_mapping}()},
\code{\link{reformat_caliber_for_ukb}()}
}
\concept{CALIBER}
