% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliber.R
\name{reformat_caliber_for_ukb}
\alias{reformat_caliber_for_ukb}
\title{Reformat and map CALIBER codes for use with UK Biobank data}
\usage{
reformat_caliber_for_ukb(
  caliber,
  all_lkps_maps,
  col_filters = codemapper::default_col_filters(),
  overlapping_disease_categories = "error",
  overlapping_disease_categories_csv = default_overlapping_disease_categories_csv()
)
}
\arguments{
\item{caliber}{A named list of data frames, created by \code{\link[=read_caliber_raw]{read_caliber_raw()}}.}

\item{all_lkps_maps}{Either a named list of lookup and mapping tables (either
data frames or \code{tbl_dbi} objects), or the path to a SQLite database
containing these tables (see also \code{\link[codemapper:build_all_lkps_maps]{build_all_lkps_maps()}} and
\code{\link[codemapper:all_lkps_maps_to_db]{all_lkps_maps_to_db()}}). If \code{NULL}, will attempt to connect to an SQLite
database named 'all_lkps_maps.db' in the current working directory, or to a
a SQLite database specified by an environmental variable named
'ALL_LKPS_MAPS_DB' (see
\href{https://resources.numbat.space/using-rprofile-and-renviron.html#renviron}{here}
for how to set environment variables using a \code{.Renviron} file). The latter
method will be used in preference.}

\item{col_filters}{A named list where each name in the list refers to the
name of a lookup or mapping table. Each item is also a named list, where
the names refer to column names in the corresponding table, and the items
are vectors of values to filter for. For example, \verb{list(my_lookup_table = list(colA = c("A", "B"))} will result in \code{my_lookup_table} being filtered
for rows where \code{colA} is either 'A' or 'B'. Uses \code{default_col_filters()} by
default. Set to \code{NULL} to remove all filters.}

\item{overlapping_disease_categories}{If 'error' (default), raises an error
if any overlapping disease categories are present after mapping. Specify
'warning' to raise a warning instead.}

\item{overlapping_disease_categories_csv}{File path to a csv containing codes
that are listed under more than one disease category within a disease. This
should have the same format as \code{\link[ukbwranglr:example_clinical_codes]{ukbwranglr::example_clinical_codes()}}, with
the author column set to 'caliber' for all rows, plus an additional 'keep'
column with 'Y' values indicating which rows to keep. By default, this is
set to \code{\link[=default_overlapping_disease_categories_csv]{default_overlapping_disease_categories_csv()}}.}
}
\value{
A named list of data frames.
}
\description{
Reformats Read 2, ICD10 and OPCS4 CALIBER codes to match the format in UK
Biobank data, and also maps from Read 2 to Read 3, as well as from ICD10 to
ICD9. See \code{vignette("caliber")} for further details.
}
\examples{
# read local copy of CALIBER repository into a named list
caliber_raw <- read_caliber_raw(dummy_caliber_dir_path())

# reformat
# reformat_caliber_for_ukb(caliber_raw, all_lkps_maps = codemapper::build_dummy_all_lkps_maps())
}
\seealso{
Other CALIBER: 
\code{\link{default_overlapping_disease_categories_csv}()},
\code{\link{download_caliber_repo}()},
\code{\link{get_caliber_categories_mapping}()},
\code{\link{read_caliber_raw}()}
}
\concept{CALIBER}
