% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{runMakeDataset}
\alias{runMakeDataset}
\title{Launch a shiny app to interactively create a UK Biobank dataset}
\usage{
runMakeDataset(
  ukb_data_dict = ukbwranglr::get_ukb_data_dict(),
  ukb_codings = ukbwranglr::get_ukb_codings(),
  options = list(launch.browser = TRUE),
  ...
)
}
\arguments{
\item{ukb_data_dict}{The UKB data dictionary (available online at the UK
Biobank
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{data
 showcase}. This should be a data frame where all columns are of type
\code{character}. By default, this is downloaded from the
\href{https://github.com/rmgpanw/ukbwranglr_resources}{\code{ukbwranglr_resources}}
 github repo.}

\item{ukb_codings}{The UKB codings dictionary. This should be a data frame
where all columns are of type \code{character}. By default, this is
downloaded from the
\href{https://github.com/rmgpanw/ukbwranglr_resources}{\code{ukbwranglr_resources}}
 github repo.}

\item{options}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment
about the ideal height/width for the app.}

\item{...}{Additional parameters passed on to \code{\link[shiny]{shinyApp}}}
}
\description{
Launches a shiny app
}
\examples{
\dontrun{
  runMakeDataset()
}
}
