% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sample_nonsig_results}
\alias{sample_nonsig_results}
\title{Take a sample of rows from non-significant results}
\usage{
sample_nonsig_results(df, pval_col = "P", sig = 5e-08, n, prop)
}
\arguments{
\item{df}{A data frame.}

\item{pval_col}{Name of column in \code{df} containing p values.}

\item{sig}{Significance threshold. P values equal to or greater than this
will be sampled.}

\item{n}{Provide either \code{n}, the number of rows, or \code{prop}, the
proportion of rows to select. If neither are supplied, \code{n = 1} will be
used.

If \code{n} is greater than the number of rows in the group (or \code{prop > 1}),
the result will be silently truncated to the group size. If the
\code{prop}ortion of a group size is not an integer, it is rounded down.}

\item{prop}{Provide either \code{n}, the number of rows, or \code{prop}, the
proportion of rows to select. If neither are supplied, \code{n = 1} will be
used.

If \code{n} is greater than the number of rows in the group (or \code{prop > 1}),
the result will be silently truncated to the group size. If the
\code{prop}ortion of a group size is not an integer, it is rounded down.}
}
\value{
A data frame.
}
\description{
May be useful when plotting Manhattan or Q-Q plots for GWAS results to speed
up plotting.
}
\examples{
dummy_manhattan_df() \%>\%
  sample_nonsig_results(pval_col = "P",
                        sig = 0.001,
                        n = 2)
}
