% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_variables.R
\name{serive_touchscreen_self_reported_med_bp_dm_chol_exog_hormones_instance_0}
\alias{serive_touchscreen_self_reported_med_bp_dm_chol_exog_hormones_instance_0}
\title{Derive indicator columns for touchscreen self-reported cholesterol, blood
pressure, insulin or exogenous hormones medications}
\usage{
serive_touchscreen_self_reported_med_bp_dm_chol_exog_hormones_instance_0(
  ukb_main,
  data_dict = ukbwranglr::get_ukb_data_dict(),
  .drop = TRUE
)
}
\arguments{
\item{ukb_main}{Main UKB dataset (data frame)}

\item{data_dict}{A data dictionary for \code{ukb_main}}

\item{.drop}{If \code{TRUE}, remove original columns from result. Default is
\code{TRUE}.}
}
\value{
A data frame with additional indicator columns for cholesterol, blood
pressure, insulin, HRT an OCP/minipill medications.
}
\description{
The data for Field IDs
\href{6177}{https://biobank.ndph.ox.ac.uk/ukb/field.cgi?id=6177} (men) and
\href{https://biobank.ndph.ox.ac.uk/ukb/field.cgi?id=6153}{6153} (women,
additionally includes values for HRT and OCP/minipill) are in wide format.
This function separates out the data in arrays 0-3 for instance 0 \emph{only}
into separate indicator 'Yes'/'No' columns, one for each medication type.
}
