% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukbwranglrextra.R
\docType{package}
\name{ukbwranglrextra}
\alias{ukbwranglrextra}
\title{ukbwranglrextra: a package for UK Biobank data wrangling}
\description{
Extra experimental functions to support the \code{ukbwranglr} R package.
}
