% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliber.R
\name{validate_overlapping_disease_categories_df}
\alias{validate_overlapping_disease_categories_df}
\title{Validation helper for \code{\link[=reformat_caliber_for_ukb]{reformat_caliber_for_ukb()}}}
\usage{
validate_overlapping_disease_categories_df(df)
}
\arguments{
\item{df}{Data frame - columns as for a clinical codes df, plus an additional
'keep' column which should only contain 'Y' and \code{NA} values.}
}
\value{
The input df, unchanged
}
\description{
Validate a user-supplied df which indicates how to handle overlapping disease
categories.
}
