% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliber.R
\name{download_caliber_repo}
\alias{download_caliber_repo}
\title{Download the caliber github repository}
\usage{
download_caliber_repo(url)
}
\arguments{
\item{url}{Download URL (e.g.
"https://github.com/spiros/chronological-map-phenotypes/archive/refs/heads/master.zip")}
}
\value{
File path to downloaded (and unzipped) repository, invisibly.
}
\description{
Downloads to \code{tempdir()} and unzips, invisibly returning the file path to the
unzipped folder.
}
\seealso{
Other CALIBER: 
\code{\link{default_overlapping_disease_categories_csv}()},
\code{\link{get_caliber_categories_mapping}()},
\code{\link{read_caliber_raw}()},
\code{\link{reformat_caliber_for_ukb}()}
}
\concept{CALIBER}
