% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_events_to_phecodes.R
\name{map_clinical_events_to_phecodes}
\alias{map_clinical_events_to_phecodes}
\title{Map UK Biobank clinical events to phecodes}
\usage{
map_clinical_events_to_phecodes(
  clinical_events,
  all_lkps_maps = NULL,
  min_date_only = FALSE
)
}
\arguments{
\item{clinical_events}{A long format data frame created by
\code{\link[ukbwranglr]{tidy_clinical_events}}, \code{\link[ukbwranglr]{tidy_gp_clinical}},
\code{\link[ukbwranglr]{tidy_gp_scripts}} or \code{\link[ukbwranglr]{make_clinical_events_db}}.
This can also be a \code{\link[dbplyr]{tbl_dbi}} object.}

\item{all_lkps_maps}{Either a named list of lookup and mapping tables (either
data frames or \code{tbl_dbi} objects), or the path to a SQLite database
containing these tables (see also \code{\link[codemapper:build_all_lkps_maps]{build_all_lkps_maps()}} and
\code{\link[codemapper:all_lkps_maps_to_db]{all_lkps_maps_to_db()}}). By default, will look for a SQLite datbase called
\code{all_lkps_maps.db} in the current working directory.}

\item{min_date_only}{If \code{TRUE}, result will be filtered for only the earliest
date per eid-phecode pair (date will be recorded as \code{NA} for cases where
there are no dates).}
}
\value{
A data frame with column names 'eid', 'source', 'index', 'code',
'icd10', 'phecode' and 'date'.
}
\description{
UK Biobank clinical events sources that are recorded in ICD10 are mapped
directly to phecodes, while non-ICD10 sources are mapped to phecodes via
ICD10.
}
\details{
Maps the following UK Biobank clinical events sources to phecodes: f40001, f40002, f20002_icd10, f40006, f41270, f40013, f41271, gpc1_r3, gpc2_r3, gpc3_r3, gpc4_r3, gpc1_r2, gpc2_r2, gpc3_r2, gpc4_r2.
}
\examples{
# dummy clinical events data frame
ukbwranglrextra:::DUMMY_CLINICAL_EVENTS

# map to phecodes
map_clinical_events_to_phecodes2(
  clinical_events = ukbwranglrextra:::DUMMY_CLINICAL_EVENTS,
  all_lkps_maps = ukbwranglrextra:::ALL_LKPS_MAPS,
  min_date_only = FALSE)
}
