% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas.R
\name{map_models}
\alias{map_models}
\title{Generate multiple models}
\usage{
map_models(
  df,
  dict,
  model_str,
  engine = logistic_reg() \%>\% set_engine("glm")
)
}
\arguments{
\item{df}{Data frame}

\item{dict}{A data frame with columns 'variable' and 'label'}

\item{model_str}{A model formula as a string, containing one variable as \code{.x}.}

\item{engine}{A \code{\link{parsnip}} engine}
}
\value{
A data frame
}
\description{
Generate multiple models
}
